change Program.main to usize, add deserialize_program function

+++ src/types/program.rs
@@ -1,4 +1,5 @@
 use crate::types::relocatable::MaybeRelocatable;
+use crate::vm::serializable_utils;
 use num_bigint::BigInt;
 
 #[derive(Clone)]
@@ -6,5 +7,5 @@ pub struct Program {
     pub builtins: Vec<String>,
     pub prime: BigInt,
     pub data: Vec<MaybeRelocatable>,
-    pub main: Option<BigInt>,
+    pub main: Option<usize>,
 }
